+++ src/types/types.ts
@@ -186,7 +186,7 @@ export interface PairHistory {
   timeframe: string;
   timeframe_ms: number;
   columns: string[];
-  data: number[];
+  data: Array<number[]>;
   length: number;
   /** Number of buy signals in this response */
   buy_signals: number;
